third_party/WebKit/LayoutTests/external/wpt/console/idlharness.any.worker-expected.txt

This is a testharness.js-based test.
PASS idl_test setup
FAIL console namespace: operation assert(boolean, any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation clear() assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation debug(any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation error(any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation info(any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation log(any...) assert_equals: operation has wrong .length expected 0 but got 1
PASS console namespace: operation table(any, [object Object])
FAIL console namespace: operation trace(any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation warn(any...) assert_equals: operation has wrong .length expected 0 but got 1
PASS console namespace: operation dir(any, object)
FAIL console namespace: operation dirxml(any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation count(DOMString) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation countReset(DOMString) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation group(any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation groupCollapsed(any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation groupEnd() assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation time(DOMString) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation timeLog(DOMString, any...) assert_equals: operation has wrong .length expected 0 but got 1
FAIL console namespace: operation timeEnd(DOMString) assert_equals: operation has wrong .length expected 0 but got 1
Harness: the test ran to completion.
